+++ src/hooks/react-query/colaboradores/colaborador.ts
@@ -4,8 +4,12 @@ import {
   convertAdvancedSearchToReactQueryKeys,
   getReactQueryPaginationKeys,
 } from '+/lib/formatters';
-import { IQueryParams } from '+/types';
-import { IAPIPaginatedResponse, IBodyRequest } from '+/types/axios';
+import { IQueryParams, IQueryByIdParams } from '+/types';
+import {
+  IAPIPaginatedResponse,
+  IAPIResponse,
+  IBodyRequest,
+} from '+/types/axios';
 import { IColaborador } from '+/types/models/colaboradores';
 
 interface Props extends IQueryParams<keyof IColaborador> {
@@ -59,3 +63,12 @@ export const useColaboradores = ({
     { refetchInterval: false }
   );
 };
+
+export const useColaboradorPorId = ({ axiosInstance, id }: IQueryByIdParams) =>
+  useQuery(['colaboradores', `id-${id}`], () =>
+    id > 0
+      ? axiosInstance
+          .get<IAPIResponse<IColaborador>>(`colaboradores/colaboradores/${id}`)
+          .then(({ data }) => data)
+      : null
+  );